requirements.txt
flask==2.3.2
tensorflow-cpu==2.16.1
pillow==10.3.0